Feature: Dedicated files for ExifUtils Methods
This commit contains EXIF based methods to transfer exif data to newly
created files, and collect and provide current EXIF data if present organized
between general and sensitive, allowing for easier viewing within the EXIF tab.

Arrays created to hold the various EXIF tags and differentiate and sort between
sensitive and general data between version >=24 and <=23

EXIF editing in the works.
