+++ app/src/main/java/net/wetfish/wetfish/utils/FileUtils.java
@@ -6,10 +6,8 @@ import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.media.ExifInterface;
 import android.media.MediaMetadataRetriever;
 import android.net.Uri;
-import android.os.Build;
 import android.preference.PreferenceManager;
 import android.provider.MediaStore;
 import android.util.Log;
@@ -38,6 +36,7 @@ public class FileUtils {
 
     private static final double UNIT_CONVERSION = 1000;
     private static final double ROUNDING_NUMBER = 100.0;
+
 
     // TODO: Might want to rename this
     public static String getAbsolutePathFromUri(Context context, Uri contentUri) {
@@ -338,216 +337,6 @@ public class FileUtils {
             return context.getString(R.string.tv_video_length_below_ten_seconds, minutes, seconds);
         } else {
             return context.getString(R.string.tv_video_length_above_ten_seconds, minutes, seconds);
-        }
-    }
-
-    public static void transferExifData(Uri originalFile, Uri newFile) {
-        Log.d("LOG_TAG", "It ranAnnananananan~~~~~~~~~~~~~~~~~~~~");
-        try {
-            // Exif attributes for android version <=23
-            String[] exifAttributesVersion23Below = new String[] {
-                    ExifInterface.TAG_APERTURE,
-                    ExifInterface.TAG_DATETIME,
-                    ExifInterface.TAG_EXPOSURE_TIME,
-                    ExifInterface.TAG_FLASH,
-                    ExifInterface.TAG_FOCAL_LENGTH,
-                    ExifInterface.TAG_GPS_ALTITUDE,
-                    ExifInterface.TAG_GPS_ALTITUDE_REF,
-                    ExifInterface.TAG_GPS_DATESTAMP,
-                    ExifInterface.TAG_GPS_LATITUDE,
-                    ExifInterface.TAG_GPS_LATITUDE_REF,
-                    ExifInterface.TAG_GPS_LONGITUDE,
-                    ExifInterface.TAG_GPS_LONGITUDE_REF,
-                    ExifInterface.TAG_GPS_PROCESSING_METHOD,
-                    ExifInterface.TAG_GPS_TIMESTAMP,
-                    ExifInterface.TAG_IMAGE_LENGTH,
-                    ExifInterface.TAG_IMAGE_WIDTH,
-                    ExifInterface.TAG_ISO,
-                    ExifInterface.TAG_MAKE,
-                    ExifInterface.TAG_MODEL,
-                    ExifInterface.TAG_ORIENTATION,
-                    ExifInterface.TAG_SUBSEC_TIME_DIGITIZED,
-                    ExifInterface.TAG_SUBSEC_TIME_ORIG,
-                    ExifInterface.TAG_WHITE_BALANCE,
-                    String.valueOf(ExifInterface.WHITEBALANCE_AUTO),
-                    String.valueOf(ExifInterface.WHITEBALANCE_MANUAL)
-               };
-
-            // Exif attributes for android version >=24
-            String[] exifAttributesVersion24Above = new String[] {
-                    ExifInterface.TAG_APERTURE_VALUE,
-                    ExifInterface.TAG_ARTIST,
-                    ExifInterface.TAG_BITS_PER_SAMPLE,
-                    ExifInterface.TAG_BRIGHTNESS_VALUE,
-                    ExifInterface.TAG_CFA_PATTERN,
-                    ExifInterface.TAG_COLOR_SPACE,
-                    ExifInterface.TAG_COMPONENTS_CONFIGURATION,
-                    ExifInterface.TAG_COMPRESSED_BITS_PER_PIXEL,
-                    ExifInterface.TAG_COMPRESSION,
-                    ExifInterface.TAG_CONTRAST,
-                    ExifInterface.TAG_COPYRIGHT,
-                    ExifInterface.TAG_CUSTOM_RENDERED,
-                    ExifInterface.TAG_DATETIME,
-                    ExifInterface.TAG_DATETIME_DIGITIZED,
-                    ExifInterface.TAG_DATETIME_ORIGINAL,
-                    ExifInterface.TAG_DEFAULT_CROP_SIZE,
-                    ExifInterface.TAG_DEVICE_SETTING_DESCRIPTION,
-                    ExifInterface.TAG_DIGITAL_ZOOM_RATIO,
-                    ExifInterface.TAG_DNG_VERSION,
-                    ExifInterface.TAG_EXIF_VERSION,
-                    ExifInterface.TAG_EXPOSURE_BIAS_VALUE,
-                    ExifInterface.TAG_EXPOSURE_INDEX,
-                    ExifInterface.TAG_EXPOSURE_MODE,
-                    ExifInterface.TAG_EXPOSURE_PROGRAM,
-                    ExifInterface.TAG_EXPOSURE_TIME,
-                    ExifInterface.TAG_FILE_SOURCE,
-                    ExifInterface.TAG_FLASH,
-                    ExifInterface.TAG_FLASHPIX_VERSION,
-                    ExifInterface.TAG_FLASH_ENERGY,
-                    ExifInterface.TAG_FOCAL_LENGTH,
-                    ExifInterface.TAG_FOCAL_LENGTH_IN_35MM_FILM,
-                    ExifInterface.TAG_FOCAL_PLANE_RESOLUTION_UNIT,
-                    ExifInterface.TAG_FOCAL_PLANE_X_RESOLUTION,
-                    ExifInterface.TAG_FOCAL_PLANE_Y_RESOLUTION,
-                    ExifInterface.TAG_GAIN_CONTROL,
-                    ExifInterface.TAG_GPS_ALTITUDE,
-                    ExifInterface.TAG_GPS_ALTITUDE_REF,
-                    ExifInterface.TAG_GPS_AREA_INFORMATION,
-                    ExifInterface.TAG_GPS_DATESTAMP,
-                    ExifInterface.TAG_GPS_DEST_BEARING,
-                    ExifInterface.TAG_GPS_DEST_BEARING_REF,
-                    ExifInterface.TAG_GPS_DEST_DISTANCE,
-                    ExifInterface.TAG_GPS_DEST_DISTANCE_REF,
-                    ExifInterface.TAG_GPS_LATITUDE,
-                    ExifInterface.TAG_GPS_LATITUDE_REF,
-                    ExifInterface.TAG_GPS_LONGITUDE,
-                    ExifInterface.TAG_GPS_LONGITUDE_REF,
-                    ExifInterface.TAG_GPS_MAP_DATUM,
-                    ExifInterface.TAG_GPS_MEASURE_MODE,
-                    ExifInterface.TAG_GPS_PROCESSING_METHOD,
-                    ExifInterface.TAG_GPS_SATELLITES,
-                    ExifInterface.TAG_GPS_SPEED,
-                    ExifInterface.TAG_GPS_SPEED_REF,
-                    ExifInterface.TAG_GPS_STATUS,
-                    ExifInterface.TAG_GPS_TIMESTAMP,
-                    ExifInterface.TAG_GPS_TRACK,
-                    ExifInterface.TAG_GPS_TRACK_REF,
-                    ExifInterface.TAG_GPS_VERSION_ID,
-                    ExifInterface.TAG_IMAGE_DESCRIPTION,
-                    ExifInterface.TAG_IMAGE_LENGTH,
-                    ExifInterface.TAG_IMAGE_UNIQUE_ID,
-                    ExifInterface.TAG_IMAGE_WIDTH,
-                    ExifInterface.TAG_INTEROPERABILITY_INDEX,
-                    ExifInterface.TAG_ISO_SPEED_RATINGS,
-                    ExifInterface.TAG_JPEG_INTERCHANGE_FORMAT,
-                    ExifInterface.TAG_JPEG_INTERCHANGE_FORMAT_LENGTH,
-                    ExifInterface.TAG_LIGHT_SOURCE,
-                    ExifInterface.TAG_MAKE,
-                    ExifInterface.TAG_MAKER_NOTE,
-                    ExifInterface.TAG_MAX_APERTURE_VALUE,
-                    ExifInterface.TAG_METERING_MODE,
-                    ExifInterface.TAG_MODEL,
-                    ExifInterface.TAG_NEW_SUBFILE_TYPE,
-                    ExifInterface.TAG_OECF,
-                    ExifInterface.TAG_ORF_ASPECT_FRAME,
-                    ExifInterface.TAG_ORF_PREVIEW_IMAGE_LENGTH,
-                    ExifInterface.TAG_ORF_PREVIEW_IMAGE_START,
-                    ExifInterface.TAG_ORF_THUMBNAIL_IMAGE,
-                    ExifInterface.TAG_ORIENTATION,
-                    ExifInterface.TAG_PHOTOMETRIC_INTERPRETATION,
-                    ExifInterface.TAG_PIXEL_X_DIMENSION,
-                    ExifInterface.TAG_PIXEL_Y_DIMENSION,
-                    ExifInterface.TAG_PLANAR_CONFIGURATION,
-                    ExifInterface.TAG_PRIMARY_CHROMATICITIES,
-                    ExifInterface.TAG_REFERENCE_BLACK_WHITE,
-                    ExifInterface.TAG_RELATED_SOUND_FILE,
-                    ExifInterface.TAG_RESOLUTION_UNIT,
-                    ExifInterface.TAG_ROWS_PER_STRIP,
-                    ExifInterface.TAG_RW2_ISO,
-                    ExifInterface.TAG_RW2_JPG_FROM_RAW,
-                    ExifInterface.TAG_RW2_SENSOR_BOTTOM_BORDER,
-                    ExifInterface.TAG_RW2_SENSOR_LEFT_BORDER,
-                    ExifInterface.TAG_RW2_SENSOR_RIGHT_BORDER,
-                    ExifInterface.TAG_RW2_SENSOR_TOP_BORDER,
-                    ExifInterface.TAG_SAMPLES_PER_PIXEL,
-                    ExifInterface.TAG_SATURATION,
-                    ExifInterface.TAG_SCENE_CAPTURE_TYPE,
-                    ExifInterface.TAG_SCENE_TYPE,
-                    ExifInterface.TAG_SENSING_METHOD,
-                    ExifInterface.TAG_SHARPNESS,
-                    ExifInterface.TAG_SHUTTER_SPEED_VALUE,
-                    ExifInterface.TAG_SOFTWARE,
-                    ExifInterface.TAG_SPATIAL_FREQUENCY_RESPONSE,
-                    ExifInterface.TAG_SPECTRAL_SENSITIVITY,
-                    ExifInterface.TAG_STRIP_BYTE_COUNTS,
-                    ExifInterface.TAG_STRIP_OFFSETS,
-                    ExifInterface.TAG_SUBFILE_TYPE,
-                    ExifInterface.TAG_SUBJECT_AREA,
-                    ExifInterface.TAG_SUBJECT_DISTANCE,
-                    ExifInterface.TAG_SUBJECT_DISTANCE_RANGE,
-                    ExifInterface.TAG_SUBJECT_LOCATION,
-                    ExifInterface.TAG_SUBSEC_TIME,
-                    ExifInterface.TAG_SUBSEC_TIME_DIG,
-                    ExifInterface.TAG_SUBSEC_TIME_ORIGINAL,
-                    ExifInterface.TAG_THUMBNAIL_IMAGE_LENGTH,
-                    ExifInterface.TAG_THUMBNAIL_IMAGE_WIDTH,
-                    ExifInterface.TAG_TRANSFER_FUNCTION,
-                    ExifInterface.TAG_USER_COMMENT,
-                    ExifInterface.TAG_WHITE_BALANCE,
-                    ExifInterface.TAG_WHITE_POINT,
-                    ExifInterface.TAG_X_RESOLUTION,
-                    ExifInterface.TAG_Y_CB_CR_COEFFICIENTS,
-                    ExifInterface.TAG_Y_CB_CR_POSITIONING,
-                    ExifInterface.TAG_Y_CB_CR_SUB_SAMPLING,
-                    ExifInterface.TAG_Y_RESOLUTION,
-                    String.valueOf(ExifInterface.WHITEBALANCE_AUTO),
-                    String.valueOf(ExifInterface.WHITEBALANCE_MANUAL)
-            };
-
-            // Check what the version of the SDK is to utilize the appropriate available ExifInterface attributes
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-
-                // Create an ExifInterface object to interact with the original file's exif data.
-                ExifInterface originalFileExif = new ExifInterface(originalFile.toString());
-
-                // Create an ExifInterface object to transfer the original file's exif data.
-                ExifInterface newFileExif = new ExifInterface(newFile.toString());
-
-                for (int i = 0; i < exifAttributesVersion24Above.length; i++) {
-                    String exifAttributeValue = originalFileExif.getAttribute(exifAttributesVersion24Above[i]);
-                    if (exifAttributeValue != null) {
-                        newFileExif.setAttribute(exifAttributesVersion24Above[i], exifAttributeValue);
-                    }
-                }
-                newFileExif.saveAttributes();
-            } else {
-                // Create an ExifInterface object to interact with the original file's exif data.
-                ExifInterface originalFileExif = new ExifInterface(originalFile.toString());
-
-                // Create an ExifInterface object to transfer the original file's exif data.
-                ExifInterface newFileExif = new ExifInterface(newFile.toString());
-
-                for (int i = 0; i < exifAttributesVersion23Below.length; i++) {
-                    String exifAttributeValue = originalFileExif.getAttribute(exifAttributesVersion23Below[i]);
-                    if (exifAttributeValue != null) {
-                        newFileExif.setAttribute(exifAttributesVersion23Below[i], exifAttributeValue);
-                    }
-                }
-                newFileExif.saveAttributes();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void writeExif(Uri absolutePath) {
-        try {
-            ExifInterface exifInterface = new ExifInterface(absolutePath.toString());
-
-            // Use exifInterface.setAttribute(ExifInterface.TAG_HERE,  String.valueOf(Whatever)
-
-        } catch (IOException e) {
-            e.printStackTrace();
         }
     }
 }
